Refactor project structure, update dependencies, and improve code quality
