+++ client/src/components/Container.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Outlet } from "react-router-dom";
 function Container(){
 return (
