+++ client/src/components/Footer.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+
 function Footer(){
 return(
     <div>
