+++ client/src/components/Header.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+
 import { FaHome } from "react-icons/fa";
 import {useNavigate} from 'react-router-dom'
 function Header(){
