+++ client/src/components/Layout.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+
 import { Outlet } from "react-router-dom";
 import Header from "./Header";
 import Footer from "./Footer";
