+++ client/src/pages/AddProperty.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import {useForm}from 'react-hook-form'
 import {useNavigate} from 'react-router-dom'
 function AddProperty(){
