+++ server/.env.example
@@ -1,0 +1,17 @@
+PORT=3000
+MONGODB_URI=mongodb://localhost:27017
+DB_NAME=realstate
+
+# Cloudinary Configuration
+CLOUDINARY_CLOUD_NAME=your_cloud_name
+CLOUDINARY_API_KEY=your_api_key
+CLOUDINARY_API_SECRET=your_api_secret
+
+# JWT Configuration
+ACCESS_TOKEN_SECRET=your_access_token_secret_here
+ACCESS_TOKEN_EXPIRY=1d
+REFRESH_TOKEN_SECRET=your_refresh_token_secret_here
+REFRESH_TOKEN_EXPIRY=10d
+
+# CORS Configuration
+CORS_ORIGIN=http://localhost:5173
