+++ server/src/index.js
@@ -2,6 +2,7 @@ import 'dotenv/config';
 import mongoose from 'mongoose';
 import app from './app.js';
 
+const PORT = process.env.PORT || 3000;
 let isConnected = false;
 
 async function connectDB() {
@@ -19,17 +20,22 @@ async function connectDB() {
   }
 }
 
-// Export a handler for Vercel
+// Start server for local development
+if (process.env.NODE_ENV !== 'production') {
+  connectDB()
+    .then(() => {
+      app.listen(PORT, () => {
+        console.log(`🚀 Server is running on port ${PORT}`);
+      });
+    })
+    .catch((error) => {
+      console.error("Failed to start server:", error);
+      process.exit(1);
+    });
+}
+
+// Export a handler for Vercel deployment
 export default async function handler(req, res) {
   await connectDB();
   return app(req, res);
 }
-
-
-
-
-
-
-
-
-
